+++ CHANGELOG.md
@@ -18,6 +18,7 @@ _Note: Gaps between patch versions are faulty, broken or test releases._
 * Added deprecation API `core/meta/deprecation`
 * Added `Object.isPromiseLike`
 * Added `Then.prototype.finally`
+* Added `core/promise/sync` API (the function `core/event/createSyncPromise` was marked as deprecated)
 
 ## v3.0.0-beta.94 (2019-12-18)